Improve resiliancy of brew installation to partial installs. Correctly check for locally installed brew not in PATH

+++ cloudbio/custom/system.py
@@ -8,7 +8,6 @@ from fabric.api import cd
 from cloudbio.custom import shared
 from cloudbio.custom.shared import _if_not_installed, _get_install, _configure_make
 
-@_if_not_installed("brew")
 def install_homebrew(env):
     """Homebrew package manager for OSX and Linuxbrew for linux systems.
 
@@ -23,6 +22,8 @@ def install_homebrew(env):
         if not env.safe_exists(brew_cmd):
             with shared._make_tmp_dir() as tmp_dir:
                 with cd(tmp_dir):
+                    if env.safe_exists("linuxbrew"):
+                        env.safe_run("rm -rf linuxbrew")
                     env.safe_run("git clone https://github.com/Homebrew/linuxbrew.git" )
                     with cd("linuxbrew"):
                         env.safe_sudo("chown %s %s" % (env.user, env.system_install))
@@ -38,10 +39,15 @@ def install_homebrew(env):
                         for path in paths:
                             if env.safe_exists("%s/%s" % (env.system_install, path)):
                                 env.safe_sudo("chown %s %s/%s" % (env.user, env.system_install, path))
-                        env.safe_run("mv Library %s" % env.system_install)
-                        env.safe_run("mv .git %s" % env.system_install)
-                        env.safe_run("mv share/man/man1/brew.1 %s/share/man/man1" % env.system_install)
-                        env.safe_run("mv bin/brew %s/bin" % env.system_install)
+                        if not env.safe_exists("%s/Library" % env.system_install):
+                            env.safe_run("mv Library %s" % env.system_install)
+                        if not env.safe_exists("%s/.git" % env.system_install):
+                            env.safe_run("mv .git %s" % env.system_install)
+                        man_dir = "share/man/man1"
+                        if not env.safe_exists("%s/%s" % (env.system_install, man_dir)):
+                            env.safe_run("mkdir -p %s/%s" % (env.system_install, man_dir))
+                        env.safe_run("mv -f %s/brew.1 %s/%s" % (man_dir, env.system_install, man_dir))
+                        env.safe_run("mv -f bin/brew %s/bin" % env.system_install)
 
 @_if_not_installed("s3fs")
 def install_s3fs(env):
